Backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
groq==0.9.0 
httpx==0.27.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
watchdog==3.0.0
python-dotenv==1.0.0
pydantic==2.5.3
matplotlib==3.8.2
plotly==5.18.0
kaleido==0.2.1
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.4.0